Moved things around - trying to layout the plugin and remember what I was doing with it
