+++ assets/js/my_script.js
@@ -1,0 +1,2 @@
+console.log( 'yes I did load' );
+
